Refactor object parsers unit tests location
Refs #37
